f89cab15-4ed5-419e-b8b0-2f3a1e189cd4/requirements.txt
discord
discord.py
discord-py-slash-command
mysqlclient
flask
redis
addict
pynacl
python-dotenv
bidict
pyyaml